+++ RELEASES.md
@@ -5,6 +5,7 @@ Released on **TBD**
 ## Breaking Changes
 - Removed CCXT adapter [#428](https://github.com/nautechsystems/nautilus_trader/issues/428).
 - Backtest configurations.
+- Renamed `UpdateOrder` to `ModifyOrder` (terminology standardization).
 
 ## Enhancements
 - Added `BacktestNode`.